Create PlaceholderOrEnrichedLinkPillView
This view lightly wraps LPLinkView, which does not have a SwiftUI
equivalent.

It is created so we can start adding support for link previews in
messages, without requiring the caller to use UIKit.

It also applies a light amount of custom styling: restricting the frame
of the view so that it appears like a "pill" instead of a full preview
to avoid taking up too much space in the message view.