Animate transition from placeholder to enriched
This commit works around a bug in SwiftUI, where applying a .animation
modifier to a Group doesn't have any effect. The transition is animated
if ZStack is used instead.

Meanwhile, make the LinkPreviewMetadata enum equatable, so that the
.animation modifier can detect changes to the value. Using the
synthesized implementation, two LinkPreviewMetadatas are equal if they
are the same case, and have payloads that are memberwise-equal (which
is the behaviour we want in this case).