Prevent vstack bubble extending to trailing edge
A spacer was used to push the time view to the right of the bubble.
However, this spacer also pushed the edge of the bubble to the trailing
edge of the screen, even if this extra space was not needed to fit the
message text in the bubble.

This did not cause a noticeable problem, since vstack was only used
when the text was long enough to spill onto multiple lines, meaning the
spacer would only cause a difference of a few pixels.

However, now we are planning to add link previews (and we want a vstack
configuration in this case even if the link is only a few characters
long), we want to constrain the width of a vstack bubble so that it is
only wide enough to hold the text.

Otherwise, a single-character message containing a link would take up
the full width of the screen, making it look out of place.