Fix a crash in the audio recorder on small screens
The maxWidth value can be small on an iPhone SE, which causes
maxSamples to be less than 0.5
And due to the casting to Int before the division, this leads
to a division by zero crash.

Remove the casting to Int, so that no longer happens.

And now that maxSamples is a CGFloat, the safeRatio workaround is no longer needed:
The early return guarantees that Double(temp.count) > maxSamples
which in turn guarantees that (Double(temp.count) / maxSamples) is greater than 1