Make Recording a struct again
This fixes a bug where the displayed time in the audio recorder doesn't update
unless the user presses one of the buttons.
Since we don't need to extend `Recording`, I opted to make it a struct again.

This is a partial revert of commit 3eee938f3b97c4, and may be squashed
with it in the next rebase.